Move blocking / unblocking logic to reusable location
